pyproject.toml
[tool.poetry]
name = "barter-desk"
version = "0.1.0"
description = ""
authors = ["Link13-tech <151656048+Link13-tech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
django-swagger-ui = "^0.1.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
